+++ data-raw/got_pov-chars-wrangle.R
@@ -2,8 +2,10 @@ library(here)
 library(jsonlite)
 library(assertthat)
 library(tidyverse)
-## devtools::install_github("jennybc/xml2@as-xml-first-try")
-## experimental as_xml() in this branch of my fork, used below
+# devtools::install_github("jennybc/xml2@as-xml-first-try")
+# experimental as_xml() in this branch of my fork, used below
+# 2022-11-11 update: branch no longer, but jenny still has locally
+# https://github.com/r-lib/xml2/issues/254
 library(xml2)
 
 ## get resource id from URL
@@ -103,10 +105,18 @@ View(pov_df)
 ## this is the basically the list that will go in the package
 got_chars <- pov_df$from_api
 
-## Delete Cersei's alias
-cersei <- which(map_chr(got_chars, "name") == "Cersei Lannister")
-got_chars[[cersei]][["aliases"]] <- list()
-
+# fixup for some NSFW language
+# if the fixup removes all vals, return empty string, to match un-fixed up data
+fixup <- function(x, key = "aliases") {
+  fixed <- str_subset(x[[key]], "[Bb]itch", negate = TRUE)
+  if (length(fixed) != length(x[[key]])) {
+    x[[key]] <- if (length(fixed) == 0) "" else fixed
+  }
+  x
+}
+got_chars <- got_chars |>
+  map(\(x) fixup(x, key = "titles")) |>
+  map(\(x) fixup(x, key = "aliases"))
 
 use_data(got_chars, overwrite = TRUE)